fix: typo rename midleware by middleware
